fix: consolidate MCP inventory package and align CLI baselines for mergeability
Fold inventory projection into registry.go, share CallTool handler registration via canonicalToolHandlers, and update CLI help/flag baselines to you-agent-factories.
