+++ pkg/cli/baseline/testdata/run_help.txt
@@ -1,6 +1,6 @@
 Load workflow and run the factory engine.
 
-For the quickest local setup, run you with no arguments. That default flow bootstraps ./factory, watches factory/inputs/task/default, keeps the runtime alive, and reports the first available dashboard URL, preferring http://localhost:7437/dashboard/ui. Default execution uses batch mode and exits after idle completion. Normal live runs record by default unless you pass --no-record. Replay artifacts are sensitive and can contain prompts, payloads, stdout, stderr, and diagnostic metadata. Use global --default-worker-model-provider and --default-worker-model to set operator-level model defaults for omitted model-worker fields. Use --continuously to keep the factory alive while idle until you cancel it. Use --with-mock-workers with an optional JSON config path to test workflows with deterministic mock worker outcomes. Use --quiet to suppress dashboard output for scripted or CI-oriented runs. Use --named with a persisted canonical factory name to resolve project-local factories before global built-ins under ~/.you-agent-factory/factories. Built-ins such as @you/tts and @you/goal materialize lazily into that global root on first use and stay editable on disk for later runs. Use --factory with a factory.json file path to run a portable factory config without guessing --dir. Selected factories can define custom invocation arguments; run you run --named <factory> --help or you run --factory <factory.json> --help to inspect signature-backed usage while keeping existing run-level flags available. In factory invocation mode, provide either trailing positional text or piped stdin text; supplying both is rejected with INVOCATION_INPUT_SOURCE_CONFLICT. Packaged @you/fusion, @you/goal, and @you/tts invocation details live in you docs packaged-fusion, you docs packaged-goal, and you docs packaged-tts. Full invocation input and return-policy details live in you docs config and you docs sessions. Use --output response-stream on supported one-shot factory invocations to render live internal session response-stream progress while the CLI owns the runtime; unsupported run shapes fall back to primary-result-only output or return INVOCATION_OUTPUT_UNSUPPORTED. Runtime logs are structured JSON rolling files grouped by UTC start date under the selected log root. Runtime metrics are a separate structured JSONL operational channel with their own rolling files and do not replace runtime logs. Environment details are record-channel diagnostics only, and system logs include command stdout/stderr only on command failures.
+For the quickest local setup, run you with no arguments. That default flow bootstraps ./factory, watches factory/inputs/task/default, keeps the runtime alive, and reports the first available dashboard URL, preferring http://localhost:7437/dashboard/ui. Default execution uses batch mode and exits after idle completion. Normal live runs record by default unless you pass --no-record. Replay artifacts are sensitive and can contain prompts, payloads, stdout, stderr, and diagnostic metadata. Use global --default-worker-model-provider and --default-worker-model to set operator-level model defaults for omitted model-worker fields. Use --continuously to keep the factory alive while idle until you cancel it. Use --with-mock-workers with an optional JSON config path to test workflows with deterministic mock worker outcomes. Use --quiet to suppress dashboard output for scripted or CI-oriented runs. Use --named with a persisted canonical factory name to resolve project-local factories before global built-ins under ~/.you-agent-factory/you-agent-factories. Built-ins such as @you/tts and @you/goal materialize lazily into that global root on first use and stay editable on disk for later runs. Use --factory with a factory.json file path to run a portable factory config without guessing --dir. Selected factories can define custom invocation arguments; run you run --named <factory> --help or you run --factory <factory.json> --help to inspect signature-backed usage while keeping existing run-level flags available. In factory invocation mode, provide either trailing positional text or piped stdin text; supplying both is rejected with INVOCATION_INPUT_SOURCE_CONFLICT. Packaged @you/fusion, @you/goal, and @you/tts invocation details live in you docs packaged-fusion, you docs packaged-goal, and you docs packaged-tts. Full invocation input and return-policy details live in you docs config and you docs sessions. Use --output response-stream on supported one-shot factory invocations to render live internal session response-stream progress while the CLI owns the runtime; unsupported run shapes fall back to primary-result-only output or return INVOCATION_OUTPUT_UNSUPPORTED. Runtime logs are structured JSON rolling files grouped by UTC start date under the selected log root. Runtime metrics are a separate structured JSONL operational channel with their own rolling files and do not replace runtime logs. Environment details are record-channel diagnostics only, and system logs include command stdout/stderr only on command failures.
 
 Usage:
   you run [flags]
@@ -29,7 +29,7 @@ Flags:
       --dir string                                                                   factory base directory (default "factory")
       --factory string                                                               path to factory.json for portable one-shot runs; use positional text or piped stdin for the invocation input
   -h, --help                                                                         help for run
-      --named string                                                                 canonical persisted factory name resolved from ./factory before ~/.you-agent-factory/factories; built-ins materialize there on first use and remain editable
+      --named string                                                                 canonical persisted factory name resolved from ./factory before ~/.you-agent-factory/you-agent-factories; built-ins materialize there on first use and remain editable
       --no-record                                                                    disable the default replay artifact for this invocation
       --output string                                                                invocation stdout mode: primary (default) or response-stream for live internal session progress on supported one-shot factory runs
       --quiet                                                                        suppress dashboard output for quiet or CI-oriented runs
